Add an interface DNSResolver for the dns resolving mechanism and a dnsCache constructor
