+++ .travis.yml
@@ -1,10 +1,13 @@
 language: go
 go:
-- 1.8
-- 1.9
-- tip
+  - 1.8
+  - 1.9
+  - 1.10
+  - 1.11
+  - 1.12
+  - tip
 matrix:
   allow_failures:
-      - go: tip
+    - go: tip
 script:
   go test -v -race -cpu=1,2,4 -bench . -benchmem ./...
